# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}},"tags":["Authentication"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}},"tags":["Authentication"]}},"/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Authentication"]}},"/file/{path}":{"get":{"operationId":"FilesConstoller_getFile","parameters":[{"name":"path","required":true,"in":"path","description":"File relative path + file name + extension","schema":{"type":"string"}}],"responses":{"default":{"description":"Returns a stream of the requested file"}},"tags":["Files"]}},"/file/upload":{"post":{"operationId":"FilesConstoller_upload","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"dest":{"type":"string"},"file":{"type":"string","format":"binary"}}}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationStatusResponse"}}}}},"tags":["Files"]}},"/file/new":{"post":{"operationId":"FilesConstoller_new","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewFileRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationStatusResponse"}}}}},"tags":["Files"]}},"/file/save":{"patch":{"operationId":"FilesConstoller_save","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveFileRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationStatusResponse"}}}}},"tags":["Files"]}},"/file/delete":{"delete":{"operationId":"FilesConstoller_delete","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewFileRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationStatusResponse"}}}}},"tags":["Files"]}},"/file/rename":{"patch":{"operationId":"FilesConstoller_rename","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RenameFileRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationStatusResponse"}}}}},"tags":["Files"]}},"/file/info/{path}":{"get":{"operationId":"FilesConstoller_info","parameters":[{"name":"path","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileInfoResponse"}}}}},"tags":["Files"]}},"/directory/root":{"get":{"operationId":"DirectoriesController_root","parameters":[],"responses":{"200":{"description":""}},"tags":["Directories"]}},"/directory/list/{path}":{"get":{"operationId":"DirectoriesController_list","parameters":[{"name":"path","required":true,"in":"path","allowEmptyValue":true,"schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DirListResponse"}}}}},"tags":["Directories"]}},"/directory/new":{"post":{"operationId":"DirectoriesController_new","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewDirRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationStatusResponse"}}}}},"tags":["Directories"]}},"/directory/delete":{"delete":{"operationId":"DirectoriesController_delete","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewDirRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationStatusResponse"}}}}},"tags":["Directories"]}},"/directory/rename":{"patch":{"operationId":"DirectoriesController_rename","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RenameDirRequest"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationStatusResponse"}}}}},"tags":["Directories"]}},"/temp/generate":{"post":{"operationId":"TempUrlConstoller_generate","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TempURLGenerateOptions"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TempURLGenerateResponse"}}}}},"tags":["Temporary URLs"]}},"/temp/{tempUrl}/get":{"get":{"operationId":"TempUrlConstoller_get","parameters":[{"name":"tempUrl","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Temporary URLs"]}},"/temp/list":{"get":{"operationId":"TempUrlConstoller_list","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TempUrl"}}}}}},"tags":["Temporary URLs"]}},"/temp/delete/{key}":{"delete":{"operationId":"TempUrlConstoller_delete","parameters":[{"name":"key","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationStatusResponse"}}}}},"tags":["Temporary URLs"]}}},"info":{"title":"Shado Cloud","description":"The Shado Cloud API description","version":"1.0","contact":{}},"tags":[{"name":"","description":""}],"servers":[],"components":{"schemas":{"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"FieldError":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"}},"required":["field","message"]},"User":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"name":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","email","name","created_at","updated_at"]},"LoginResponse":{"type":"object","properties":{"errors":{"type":"array","items":{"$ref":"#/components/schemas/FieldError"}},"user":{"$ref":"#/components/schemas/User"}},"required":["errors","user"]},"RegisterRequest":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["name","email","password"]},"OperationStatusResponse":{"type":"object","properties":{"errors":{"type":"array","items":{"$ref":"#/components/schemas/FieldError"}},"status":{"type":"string","enum":["FAILED","SUCCESS","ONGOING","UNKNOWN"]}},"required":["errors","status"]},"NewFileRequest":{"type":"object","properties":{"name":{"type":"string","example":"Relative path + file name + file extension"}},"required":["name"]},"SaveFileRequest":{"type":"object","properties":{"name":{"type":"string","example":"Relative path + file name + file extension"},"content":{"type":"string"},"append":{"type":"boolean","default":false}},"required":["name","content","append"]},"RenameFileRequest":{"type":"object","properties":{"name":{"type":"string","example":"Relative path + file name + file extension"},"newName":{"type":"string","example":"Name + extension"}},"required":["name","newName"]},"FileInfo":{"type":"object","properties":{"name":{"type":"string"},"extension":{"type":"string"},"path":{"type":"string"},"mime":{"type":"string"},"size":{"type":"number"},"is_image":{"type":"boolean"},"is_video":{"type":"boolean"},"is_pdf":{"type":"boolean"},"is_text":{"type":"boolean"},"is_audio":{"type":"boolean"}},"required":["name","extension","path","mime","size","is_image","is_video","is_pdf","is_text","is_audio"]},"FileInfoResponse":{"type":"object","properties":{"errors":{"type":"array","items":{"$ref":"#/components/schemas/FieldError"}},"status":{"type":"string","enum":["FAILED","SUCCESS","ONGOING","UNKNOWN"]},"data":{"$ref":"#/components/schemas/FileInfo"}},"required":["errors","status","data"]},"DirListResponse":{"type":"object","properties":{"errors":{"type":"array","items":{"$ref":"#/components/schemas/FieldError"}},"status":{"type":"string","enum":["FAILED","SUCCESS","ONGOING","UNKNOWN"]},"parent":{"type":"string"},"data":{"type":"array","items":{"type":"object"}}},"required":["errors","status","parent","data"]},"NewDirRequest":{"type":"object","properties":{"name":{"type":"string","example":"relative path + name"}},"required":["name"]},"RenameDirRequest":{"type":"object","properties":{"name":{"type":"string","example":"relative path + name"},"newName":{"type":"string","example":"new relative path + name"}},"required":["name","newName"]},"TempURLGenerateOptions":{"type":"object","properties":{"max_requests":{"type":"number","default":100},"is_readonly":{"type":"boolean","default":true},"expires_at":{"format":"date-time","type":"string","default":"2022-01-16T07:24:47.480Z"},"filepath":{"type":"string"}},"required":["max_requests","is_readonly","expires_at","filepath"]},"TempURLGenerateResponse":{"type":"object","properties":{"url":{"type":"string"}},"required":["url"]},"TempUrl":{"type":"object","properties":{"id":{"type":"number"},"url":{"type":"string"},"filepath":{"type":"string"},"requests":{"type":"number"},"max_requests":{"type":"number"},"is_readonly":{"type":"boolean"},"expires_at":{"format":"date-time","type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"is_valid":{"type":"boolean"}},"required":["id","url","filepath","requests","max_requests","is_readonly","expires_at","created_at","updated_at","is_valid"]}}}}